requirements.txt
Flask==3.1.0
flask_swagger_ui==4.11.1
python-dotenv==1.0.1
